+++ js/toggle-modal.js
@@ -1,0 +1,13 @@
+import { bigPicture } from './info-modal.js';
+
+const openModal = () => {
+  bigPicture.classList.remove('hidden');
+  document.body.classList.add('modal-open');
+};
+
+const closeModal = () => {
+  bigPicture.classList.add('hidden');
+  document.body.classList.remove('modal-open');
+};
+
+export { closeModal, openModal };
